Lots o' changes
Added in a bunch of pictures for the info page. Put those pictures in the info page. Unfortunately, the pictures are all formatted weird and the labels no longer show up over the pictures, so this needs to be fixed. Also edited the home page to newer versions of Forestr description & motivation.